+++ 727 server/src/com/rs/net/encoders/WorldPacketsEncoder.java
@@ -310,7 +310,7 @@ public class WorldPacketsEncoder extends Encoder {
 	public void sendRunEnergy() {
 		OutputStream stream = new OutputStream(2);
 		stream.writePacket(player, 64);
-		stream.writeByte(player.getRunEnergy());
+		stream.writeByte((int) player.getRunEnergy());
 		session.write(stream);
 	}
